FEJ: moved exploiter utils from pipeline.py to utils.py
